tests/wpt/tests/html/canvas/element/webgpu-access/2d.webgpuaccess.getTextureFormat.rgba16f.tentative.https.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.webgpuaccess.getTextureFormat.rgba16f.tentative.https</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.webgpuaccess.getTextureFormat.rgba16f.tentative.https</h1>
<p class="desc">getTextureFormat() returns RGBA16F for a float16 context</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("getTextureFormat() returns RGBA16F for a float16 context");
_addTest(function(canvas, ctx) {

  _assertSame(ctx.getTextureFormat(), "rgba16float", "ctx.getTextureFormat()", "\"rgba16float\"");

}, {colorSpace: "display-p3", pixelFormat: "float16"});
</script>
